cli/command/image: un-export RunSave
This function was exported in e43c7920ea950b0c8f2004bce114e71f7f993854
for use of "docker app", which is now deprecated. The signature of this
function also depended on a non-exported type, so it could not be used
externally.

Make it internal again, as it was never designed to be exported. There
are no known external consumers of this function.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>