cli/command/image: deprecate RunPull and make internal
This function was exported in 812f1136850f6c18bbe3f1b2a960a8ff8a8413f3
for use in other parts of the CLI, but it's now only used locally.

Make it internal again, as it was never designed to be exported. There
are no known external consumers of this function, but deprecating it
first, in case there are.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>