Use `DOCKER_AUTH_CONFIG` env as credential store
This patch gives the CLI the ability to pick up the `DOCKER_AUTH_CONFIG`
environment variable and use it as a credential store.

Credentials stored in `DOCKER_AUTH_CONFIG` would take precedence over any
credentials in `~/.docker/config.json` or stored in the credential helper.

Destructive actions, such as deleting a credential would result in a noop if
found in the environment credential. Credentials found in the file or
native store would get removed.

Signed-off-by: Alano Terblanche <18033717+Benehiko@users.noreply.github.com>